+++ src/cmd.cpp
@@ -723,16 +723,27 @@ static int builtin_hash(DSState &state, ArrayObject &argvObj) {
 
 static StrRefMap<CodeCompOp> initCompActions() {
   return {
-      {"file", CodeCompOp::FILE},       {"dir", CodeCompOp::DIR},
-      {"module", CodeCompOp::MODULE},   {"exec", CodeCompOp::EXEC},
-      {"tilde", CodeCompOp::TILDE},     {"command", CodeCompOp::COMMAND},
-      {"cmd", CodeCompOp::COMMAND},     {"external", CodeCompOp::EXTERNAL},
-      {"builtin", CodeCompOp::BUILTIN}, {"udc", CodeCompOp::UDC},
-      {"dyna", CodeCompOp::DYNA_UDC},   {"variable", CodeCompOp::VAR},
-      {"var", CodeCompOp::VAR},         {"env", CodeCompOp::ENV},
-      {"signal", CodeCompOp::SIGNAL},   {"user", CodeCompOp::USER},
-      {"group", CodeCompOp::GROUP},     {"stmt_kw", CodeCompOp::STMT_KW},
-      {"expr_kw", CodeCompOp::EXPR_KW}, {"type", CodeCompOp::TYPE},
+      {"file", CodeCompOp::FILE},
+      {"dir", CodeCompOp::DIR},
+      {"module", CodeCompOp::MODULE},
+      {"exec", CodeCompOp::EXEC},
+      {"tilde", CodeCompOp::TILDE},
+      {"command", CodeCompOp::COMMAND},
+      {"cmd", CodeCompOp::COMMAND},
+      {"external", CodeCompOp::EXTERNAL},
+      {"builtin", CodeCompOp::BUILTIN},
+      {"udc", CodeCompOp::UDC},
+      {"dyna", CodeCompOp::DYNA_UDC},
+      {"variable", CodeCompOp::VAR},
+      {"var", CodeCompOp::VAR},
+      {"env", CodeCompOp::ENV},
+      {"valid_env", CodeCompOp::VALID_ENV},
+      {"signal", CodeCompOp::SIGNAL},
+      {"user", CodeCompOp::USER},
+      {"group", CodeCompOp::GROUP},
+      {"stmt_kw", CodeCompOp::STMT_KW},
+      {"expr_kw", CodeCompOp::EXPR_KW},
+      {"type", CodeCompOp::TYPE},
   };
 }